Struct vectormath::quaternion::Quat

pub struct Quat<T> {
    pub w: T,
    pub x: T,
    pub y: T,
    pub z: T,
}

Quaternion

Definition

$$ \mathbf{q} = w + x \mathbf{i} + y \mathbf{j} + z \mathbf{k} = w + \mathbf{v}, \\ \textrm{where } w \in \mathbb{R}, \mathbf{v} \in \mathbb{R}^3 \\ $$

Basis

$$ \mathbf{i}^2 = \mathbf{j}^2 = \mathbf{k}^2 = \mathbf{i} \mathbf{j} \mathbf{k} = -1, \\ \textrm{where } \mathbf{i}, \mathbf{j}, \mathbf{k} \textrm{ are basis elements} $$

Set

$$ \textrm{the set of } \mathbf{q} \textrm{ is denoted } \mathbb{H} $$

Fields

w

real part (scalar part)

x

1st imaginary part (x of vector part)

y

2nd imaginary part (y of vector part)

z

3rd imaginary part (z of vector part)

Methods

impl<T: Float> Quat<T>

fn conj(self) -> Quat<T>

Conjugate

Definition

$$ \mathbf{q}^\ast = w - x \mathbf{i} - y \mathbf{j} - z \mathbf{k} $$

fn norm(self) -> T

Magnitude

Definition

$$ |\mathbf{q}| = \sqrt{\mathbf{q} \mathbf{q}^\ast} $$

fn norm_squared(self) -> T

Squared Magnitude

Definition

$$ |\mathbf{q}|^2 = \mathbf{q} \mathbf{q}^\ast = w^2 + \mathbf{v} \cdot \mathbf{v} = w^2 + x^2 + y^2 + z^2 $$

fn normalize(self) -> Quat<T>

Normalize

Definition

$$ \hat{\mathbf{q}} = \frac{\mathbf{q}}{|\mathbf{q}|} $$

fn dot(self, rhs: Quat<T>) -> T

Dot Product

Definition

$$ \mathbf{q}_1 \cdot \mathbf{q}_2 = w_1 w_2 + \mathbf{v}_1 \cdot \mathbf{v}_2 $$

fn rotate(self, v: Vec3<T>) -> Vec3<T>

Transform a vector with the quaternion

Definition

$$ \mathbf{p}' = \mathbf{q} \mathbf{p} \mathbf{q}^{-1}, \\ \textrm{where } \mathbf{p} = 0 + x \mathbf{i} + y \mathbf{j} + z \mathbf{k} $$

fn lerp(self, dest: Quat<T>, t: T) -> Quat<T>

linear interpolation

Definition

$$ \mathbf{q} = \mathrm{lerp}(\mathbf{q}_0, \mathbf{q}_1; t) = \mathbf{q}_0 + t (\mathbf{q}_1 - \mathbf{q}_0), \\ \textrm{where } t \in \mathbb{R} $$

fn slerp(self, dest: Quat<T>, t: T) -> Quat<T>

spherical linear interpolation

Definition

$$ \mathbf{q} = \mathrm{slerp}(\mathbf{q}_0, \mathbf{q}_1; t) = \frac{ \mathbf{q}_0 \sin ((1 - t) \phi) + \mathbf{q}_1 \sin (t \phi) }{ \sin \phi }, \\ \textrm{where } \cos \phi = \cos \frac{\theta}{2} = w_0 w_1 + x_0 x_1 + y_0 y_1 + z_0 z_1 $$

See Also

Martin Baker

fn xyz(self) -> Vec3<T>

extract xyz components

impl<T: Float> Quat<T>

Constructors

fn identity() -> Quat<T>

Construct an identity quaternion

fn zero() -> Quat<T>

Construct a zero quaternion

fn from_angle_axis(angle: Rad<T>, axis: Vec3<T>) -> Quat<T>

Construct a quaternion that rotates by angle around axis

fn from_vectors(v0: Vec3<T>, v1: Vec3<T>) -> Quat<T>

Construct a quaternion that rotates from one vector to another